EFRS-1025: move ConsistenceController to core-app; no auth for consistency check;
