--- java/admin/src/main/java/com/exadel/frs/dto/VersionConsistenseDto.java
@@ -1,12 +1,0 @@
-package com.exadel.frs.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class VersionConsistenseDto {
-    boolean demoFaceCollectionIsInconsistent;
-    boolean dbIsInconsistent;
-    boolean saveImagesToDB;
-}
