+++ java/api/src/main/java/com/exadel/frs/core/trainservice/filter/SecurityValidationFilter.java
@@ -83,7 +83,7 @@ public class SecurityValidationFilter implements Filter {
         val httpResponse = (HttpServletResponse) servletResponse;
 
         String requestURI = httpRequest.getRequestURI();
-        if (!requestURI.matches("^/(swagger|webjars|v2|api/v1/migrate).*$")) {
+        if (!requestURI.matches("^/(swagger|webjars|v2|api/v1/migrate|consistence/status).*$")) {
             val headersMap =
                     list(httpRequest.getHeaderNames()).stream()
                             .collect(Collectors.<String, String, List<String>>toMap(
